backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-filter==25.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
dj-database-url==2.2.0
psycopg[binary]==3.2.6
mysqlclient==2.2.7
python-decouple==3.8
drf-spectacular==0.28.0
gunicorn==23.0.0
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
fpdf2==2.8.2
openpyxl==3.1.5
Pillow==11.1.0
